Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -13,7 +13,7 @@ int main() {
         if (b==0) {
             printf("Error\n");
         } else {
-            printf("%d",a/b);
+            printf("%.0f",(float)a/b);
         }
     } else {
         printf("Error\n");
